+++ app/src/main/java/org/simple/clinic/home/overdue/OverdueAppointment.kt
@@ -159,5 +159,23 @@ data class OverdueAppointment(
         facilityUuid: UUID,
         scheduledBefore: LocalDate
     ): PagingSource<Int, OverdueAppointment>
+
+    @Query("""
+      $OVERDUE_APPOINTMENTS_QUERY
+      WHERE
+        (
+          P.uuid IN (SELECT uuid FROM PatientFts WHERE fullName MATCH "*"||:query||"*") OR
+          P.addressUuid IN (SELECT uuid FROM PatientAddressFts WHERE colonyOrVillage MATCH "*"||:query||"*")
+        ) AND
+        IFNULL(patientAssignedFacilityUuid, appt_facilityUuid) = :facilityUuid AND
+        appt_scheduledDate < :scheduledBefore
+      GROUP BY appt_patientUuid
+      ORDER BY fullName COLLATE NOCASE
+    """)
+    fun searchImmediate(
+        query: String,
+        facilityUuid: UUID,
+        scheduledBefore: LocalDate
+    ): List<OverdueAppointment>
   }
 }