+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 - Add support for selecting all overdue appointment search results
 - Add support for runtime permissions and network status in overdue search screen
 - Implement `AppointmentRepository#searchOverduePatientsImmediate`
+- Fix search results appointments not getting fully downloaded/shared/selected
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ### Changes