+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 - Implement `AppointmentRepository#searchOverduePatientsImmediate`
 - Fix search results appointments not getting fully downloaded/shared/selected
 - Remove unused code from overdue search
+- Add overdue patients select and download feature flag key
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ### Changes