+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Remove unused code from overdue search
 - Add overdue patients select and download feature flag key
 - Fix overdue search not displaying no patients found error view
+- Point accordion downwards when overdue list is expanded
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ### Fixes