+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Add overdue patients select and download feature flag key
 - Fix overdue search not displaying no patients found error view
 - Point accordion downwards when overdue list is expanded
+- Fix `ContactPatientBottomSheet` not opening completely
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ### Fixes