+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Point accordion downwards when overdue list is expanded
 - Fix `ContactPatientBottomSheet` not opening completely
+- Update select and download feature flag key to v2
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-07-25-8342