+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Point accordion downwards when overdue list is expanded
 - Fix `ContactPatientBottomSheet` not opening completely
 - Update select and download feature flag key to v2
+- Bump Sentry to v6.3.0
+- Bump Sentry gradle plugin to v3.1.3
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-07-25-8342